Moved normalize_proba from preprocessing to utils/extmath
